Update to v0.7.0
Updated the readme.
Added a force reset to the Console class's sub-objects.
Added the 'reset' flag to the Color interface.
Added the ControlSequenceIntroducer (CSI) class.
Updated docblock comments.
If the ProgressBar Rainbow256Mode flag is set, it will use it regardless of the RainbowMode flag. (They are now independent of each other.)
Round the rate of the ProgressBar to 2.
Allow tables to have no border.
Added screen buffer switching.
Added some tests.